# internal/swagger/api_v1.0.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "GTB Swagger",
      "description": "GTB server API",
      "contact": {
        "email": "developer@energie-ip.com"
      },
      "version": "1.0"
    },
    "servers": [
      {
        "url": "/{basePath}",
        "variables": {
          "basePath": {
            "default": "v1.0"
          }
        }
      }
    ],
    "paths": {
...
      "/setup/group/{groupID}": {
        "get": {
          "tags": [
            "setup"
          ],
          "summary": "getGroupSetup",
          "description": "Return Group Setup configuration",
          "operationId": "GetGroupSetup",
          "parameters": [
            {
              "name": "groupID",
              "in": "path",
              "description": "Group Number",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {},
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GroupSetup"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        },
        "delete": {
          "tags": [
            "setup"
          ],
          "summary": "removeGroupSetup",
          "description": "Remove Group Setup configuration",
          "operationId": "RemoveGroupSetup",
          "parameters": [
            {
              "name": "groupID",
              "in": "path",
              "description": "Group Number",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {}
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
      "/setup/group": {
        "post": {
          "tags": [
            "setup"
          ],
          "summary": "setGroupSetup",
          "description": "Group Setup configuration",
          "operationId": "SetGroupSetup",
          "parameters": [],
          "requestBody": {
            "description": "Group config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupSetup"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {}
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
      "/config/group": {
        "post": {
          "tags": [
            "config"
          ],
          "summary": "setGroupConfig",
          "description": "Group runtime configuration",
          "operationId": "SetGroupConfig",
          "parameters": [],
          "requestBody": {
            "description": "Group config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupConfig"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {}
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
      "/command/group": {
        "post": {
          "tags": [
            "command"
          ],
          "summary": "sendGroupCommand",
          "description": "Send Group command",
          "operationId": "SendGroupCommand",
          "parameters": [],
          "requestBody": {
            "description": "Group command",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupCommand"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {}
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
      "/status/groups": {
        "get": {
          "tags": [
            "status"
          ],
          "summary": "getGroupsStatus",
          "description": "Return all group status",
          "operationId": "GetGroupsStatus",
          "parameters": [],
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {},
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/GroupStatus"
                    },
                    "description": "sucessful operation"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
      "/status/group/{groupID}": {
        "get": {
          "tags": [
            "status"
          ],
          "summary": "getGroupStatus",
          "description": "Return the given group status",
          "operationId": "GetGroupStatus",
          "parameters": [
            {
              "name": "groupID",
              "in": "path",
              "description": "group ID",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {},
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GroupStatus"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "GroupStatus": {
          "title": "GroupStatus",
          "required": [
            "group"
          ],
          "type": "object",
          "properties": {
            "group": {
              "type": "integer",
              "description": "Associate group",
              "format": "int32"
            },
            "auto": {
              "type": "boolean",
              "description": "Automatic mode"
            },
            "watchdog": {
              "type": "integer",
              "description": "watchdog in second",
              "format": "int32"
            },
            "slopeStartManual": {
              "type": "integer",
              "description": "slopeStart Manual in second",
              "format": "int32"
            },
            "slopeStopManual": {
              "type": "integer",
              "description": "slopeStop Manual in second",
              "format": "int32"
            },
            "slopeStartAuto": {
              "type": "integer",
              "description": "slopeStart Auto in second",
              "format": "int32"
            },
            "slopeStopAuto": {
              "type": "integer",
              "description": "slopeStop Auto in second",
              "format": "int32"
            },
            "sensorRule": {
              "$ref": "#/components/schemas/SensorRule"
            },
            "correctionInterval": {
              "type": "integer",
              "description": "correctionInterval in second",
              "format": "int32"
            },
            "leds": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Array of LED mac addresses"
            },
            "sensors": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Array of Sensor mac addresses"
            },
            "ruleBrightness": {
              "type": "integer",
              "description": "Brightness in Lux",
              "format": "int32"
            },
            "firstDayOffset": {
              "type": "integer",
              "description": "Offset in percentage between 1st and 2nd Days",
              "format": "int32"
            },
            "rulePresence": {
              "type": "integer",
              "description": "Presence in Second",
              "format": "int32"
            },
            "error": {
              "type": "integer",
              "description": "Error code value: 0 means no error",
              "format": "int32"
            },
            "presence": {
              "type": "boolean",
              "description": "is set to true if at least one sensor detect a movement"
            },
            "brightness": {
              "type": "integer",
              "description": "Compute Brightness in the group (Lux)",
              "format": "int32"
            },
            "temperature": {
              "type": "integer",
              "description": "Compute Temperature in the group (1/10°C)",
              "format": "int32"
            },
            "humidity": {
              "type": "integer",
              "description": "Compute Humidity in the group (1/10%)",
              "format": "int32"
            },
            "windowsOpened": {
              "type": "boolean",
              "description": "is set to true if at least one window is opened"
            },
            "timeToAuto": {
              "type": "integer",
              "description": "Remaining time before switching back to auto mode",
              "format": "int32"
            },
            "timeToLeave": {
              "type": "integer",
              "description": "Remaining time before considering the room as empty",
              "format": "int32"
            },
            "setpointLeds": {
              "type": "integer",
              "description": "Current led setpoint value in percentage",
              "format": "int32"
            },
            "firstDay": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Array of LED mac addresses which are in 1st Day"
            },
            "setpointLedsFirstDay": {
              "type": "integer",
              "description": "Current led setpoint value in percentage for the 1st Day",
              "format": "int32"
            },
            "friendlyName": {
              "type": "string",
              "description": "Sensor FriendlyName"
            }
          }
        },
        "GroupSetup": {
          "title": "GroupSetup",
          "required": [
            "group"
          ],
          "type": "object",
          "properties": {
            "group": {
              "type": "integer",
              "description": "Associate group",
              "format": "int32"
            },
            "auto": {
              "type": "boolean",
              "description": "Automatic mode"
            },
            "watchdog": {
              "type": "integer",
              "description": "watchdog in second",
              "format": "int32"
            },
            "slopeStartManual": {
              "type": "integer",
              "description": "slopeStart Manual in second",
              "format": "int32"
            },
            "slopeStopManual": {
              "type": "integer",
              "description": "slopeStop Manual in second",
              "format": "int32"
            },
            "slopeStartAuto": {
              "type": "integer",
              "description": "slopeStart Auto in second",
              "format": "int32"
            },
            "slopeStopAuto": {
              "type": "integer",
              "description": "slopeStop Auto in second",
              "format": "int32"
            },
            "sensorRule": {
              "$ref": "#/components/schemas/SensorRule"
            },
            "correctionInterval": {
              "type": "integer",
              "description": "correctionInterval in second",
              "format": "int32"
            },
            "leds": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Array of LED mac addresses"
            },
            "sensors": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Array of Sensor mac addresses"
            },
            "ruleBrightness": {
              "type": "integer",
              "description": "Brightness in Lux",
              "format": "int32"
            },
            "rulePresence": {
              "type": "integer",
              "description": "Presence in Second",
              "format": "int32"
            },
            "friendlyName": {
              "type": "string",
              "description": "Sensor FriendlyName"
            },
            "firstDay": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Array of LED mac addresses which are in 1st Day"
            },
            "firstDayOffset": {
              "type": "integer",
              "description": "Offset in percentage between 1st and 2nd Days",
              "format": "int32"
            }
          }
        },
        "GroupConfig": {
          "title": "GroupConfig",
          "required": [
            "group"
          ],
          "type": "object",
          "properties": {
            "group": {
              "type": "integer",
              "description": "Associate group",
              "format": "int32"
            },
            "auto": {
              "type": "boolean",
              "description": "Automatic mode"
            },
            "watchdog": {
              "type": "integer",
              "description": "watchdog in second",
              "format": "int32"
            },
            "slopeStartAuto": {
              "type": "integer",
              "description": "slopeStart Auto in second",
              "format": "int32"
            },
            "slopeStopAuto": {
              "type": "integer",
              "description": "slopeStop Auto in second",
              "format": "int32"
            },
            "slopeStartManual": {
              "type": "integer",
              "description": "slopeStart Manual in second",
              "format": "int32"
            },
            "slopeStopManual": {
              "type": "integer",
              "description": "slopeStop Manual in second",
              "format": "int32"
            },
            "sensorRule": {
              "$ref": "#/components/schemas/SensorRule"
            },
            "correctionInterval": {
              "type": "integer",
              "description": "correctionInterval in second",
              "format": "int32"
            },
            "leds": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Array of LED mac addresses"
            },
            "sensors": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Array of Sensor mac addresses"
            },
            "ruleBrightness": {
              "type": "integer",
              "description": "Brightness in Lux",
              "format": "int32"
            },
            "rulePresence": {
              "type": "integer",
              "description": "Presence in Second",
              "format": "int32"
            },
            "friendlyName": {
              "type": "string",
              "description": "Sensor FriendlyName"
            },
            "firstDay": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Array of LED mac addresses which are in 1st Day"
            },
            "firstDayOffset": {
              "type": "integer",
              "description": "Offset in percentage between 1st and 2nd Days",
              "format": "int32"
            }
          }
        },
        "GroupCommand": {
          "title": "GroupCommand",
          "required": [
            "group"
          ],
          "type": "object",
          "properties": {
            "group": {
              "type": "integer",
              "description": "Associate group",
              "format": "int32"
            },
            "auto": {
              "type": "boolean",
              "description": "Automatic mode"
            },
            "setpointLeds": {
              "type": "integer",
              "description": "Led Setpoint value",
              "format": "int32"
            },
            "setpointBlinds": {
              "type": "integer",
              "description": "Blind Setoint value",
              "format": "int32"
            },
            "setpointSlats": {
              "type": "integer",
              "description": "Slat Setoint value",
              "format": "int32"
            }
          }
        },
...
        "Error": {
          "title": "Error",
          "required": [
            "code",
            "message"
          ],
          "type": "object",
          "properties": {
            "code": {
              "type": "integer",
              "format": "int32"
            },
            "message": {
              "type": "string"
            }
          }
        },
...
        "SensorRule": {
          "title": "SensorRule",
          "enum": [
            "average",
            "min",
            "max"
          ],
          "type": "string",
          "description": "Rule when multiple sensors are presents in the same group",
          "example": "average"
        },
...
      },
      "securitySchemes": {
        "Authorization": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT",
            "description": "For accessing the API a valid JWT token must be passed in all the queries in the Authorization header. Authorization header : Bearer xxxxxx.yyyyyyy.zzzzzz"
        }
      }
    },
    "security": [
      {
        "Authorization": []
      }
    ],
...
  }